[etop] Extend timer when fetching process info from node
Originally, etop would only wait for 1 second after starting the
collection of process info from the observed node. If data was not
received in time, etop would exit with reason 'connection_lost'.

In some cases, however, this timer would expire due to the amount of
processes on the observed node. To avoid this, the timer is now
extended to the same as the update interval, which means that if the
problem occurs, it can be helped by extending the interval time.