Address timing attack in cookies
It was recently disclosed that Erlang cookies are
vulnerable to timing attacks in case an attacker
can access a machine and the machine accepts node
connections without TLS.

This patch removes the timing attack by no longer
converting the cookie to an atom. This also means
the cookie will no longer be listed whenever atoms
are listed (although it will still be dumped with
the auth process heap).

TODO:

  * Introduce a new API for reading and setting cookies and update the docs
  * Remove the deprecated auth API
  * Revisit auth.erl old messages