Add a new SSL option; client_cert_verify_fun
TLS client certificates work by, once the client key exchange in the TLS
handshake completes, the client can send proof it has access to the key
described by the client certificate. It does this by signing a hash of
all the handshake messages sent up to that point on the TLS session. See
RFC 5246 Section 7.4.8 for more details.

This patch allows the user to specify a function to do the signing,
rather than requiring the user to supply the key directly. This is
useful when using a hardware security module, where a key is stored in
such a way that it can not be read out directly, data must be fed into
the HSM to be signed, and the resulting signature is the output.