compile_SUITE: Don't create a directory called 'core'
A core dump can't be created with the name 'core' if there is
a directory named 'core'.

Rename the test case from core/1 to core_pp/1 so that the directory
name can be the same as the test case name. It also makes sense
to use a less generic name for the test case.