Fix httpc timeout for redirects
Probably since 6153ba7 (OTP R13B04) the httpc timeout setting does not
work for redirects (when autoredirect is true). With this patch a new
timer is started for the new (redirected) requests. This means that a
simple redirected request could return after 2*timeout milliseconds.

This is the first part to fix https://bugs.erlang.org/browse/ERL-420