stdlib: Limit the size of gen_server's error events
The state of the gen_server is limited in error events before exiting
(if the Kernel variable error_logger_format_depth is set).

An alternative is to let the error_logger limit all messages
(error_msg, format, warning_msg, info_msg), which would not limit
reports and also add a smallish overhead to event logging.

It is not decided if the alternative is to be implemented.