kernel: Add error_logger:limit_term/1
Calling error_logger:limit_term/1 before sending terms as events to
the error_logger can be used for limiting the size of the messages.
Doing so will not change the output, but potentially save memory.

The Kernel variable error_logger_format_depth is used when limiting
the size of terms.