stdlib: Limit the size of proc_lib's crash reports
The size of the message queue and the dictionary is limited in
crash reports.

To avoid creating the potentially huge list of messages of the message
queue, messages are received (if the Kernel variable
error_logger_format_depth is set).

The tag 'message_queue_len' has been added to the crash report.