proc_lib: Don't call erlang:get_stacktrace/0 twice
proc_lib calls erlang:get_stacktrace/0 twice, which is unnecessary,
and potentially unsafe since there are calls to many functions
in between. Any of the calls could potentially cause and catch
an exception, invalidating the stacktrace.

Only call erlang:get_stacktrace/0 once, and pass the result to
the second place where it is needed.