Fix non-delivery of ssl_closed message in active once
The commit 8b10920 (OTP 19.3.1) fixed the non-delivery of final TLS
record in {active, once}, but this causes the ssl_closed message to be
lost when the TCP connection closes before ssl:close/1. The patch
restores the behavior of OTP 18.

This is the second part to fix https://bugs.erlang.org/browse/ERL-420