erts: Add undocumented option to do default SIGTERM
This is a stopgap measure before the release of OTP-20 where
that makes it possible for the user to make the vm not do
anything with SIGTERM and instead rely on the OS default.

To enable this behaviour the user should set the environment
variable ERL_ZZ_SIGTERM_KILL="true".