Add stacktrace entries to BIF calls from emulator
The goal of this change is to improve debugging of
emulator calls. For example, the following code

    rem(1, y)

will error with atom `badarith` when y is 0 and the
stacktrace has no entry for `erlang:rem/2`, making
such cases very hard to debug. This patch makes it
so the stacktrace includes `erlang:rem(1, 0)`.

The following emulator BIFs have been changed:

  * band/2
  * bnot/1
  * bor/2
  * bsl/2
  * bsr/2
  * bxor/2
  * div/2
  * element/2
  * int_div/2
  * rem/2
  * sminus/2
  * splus/2
  * stimes/2