Make sure that asynchronous replies are not lost
If an synchronous GC was requested by calling:

     erlang:garbage_collect(Pid, [{async,Ref}])

the reply message could in certain circumstances be
lost.

The problem is in cleanup_sys_tasks() in erl_process.c.
If there were at least one dirty task, only the first
dirty task would be cleaned up. All other systems tasks
would not be cleaned up (that is, no replies would be
sent to other processes waiting for the tasks to finish).