Fix process_SUITE:low_prio/1
It is expected that low-priority processes should be allowed
to run about 1/8 of the time of normal-priority processes.

The proportion calculated was usually considerably lower than
1/8. It seems that the main reason is the punishement in
reductions when sending to a process that alreday has many
messages in the message queue. I have verified that by running
the emulator with the +snsp to turn off the send punishment.

To avoid that punishment, let the server process run at high
priority so that it can keep its message queue short.

It seems that the proportion is now very close to 1/8. Therefore,
we can tighten the margins for the proportion.