Update sasl_report_SUITE to accept new size of sasl.log
PR-1448 caused sasl_report_SUITE:gen_server_crash to produce a
slightly bigger sasl.log. This is because the Reason used in the crash
report is changed from {ExitReason,Stacktrace} to ExitReason only,
where ExitReason={badmatch,BigGbSet}. The term is written with a depth
limit of 30, and when removing the top tuple, a new level of the
gb_set is exposed. This level contains more data than the removed
Stacktrace, and thus the resulting log gets a bit bigger than before.