Make sure that asynchronous replies are not lost
(First attempt to fix in 23f132d9ab776a.)

If an synchronous GC was requested by calling:

    erlang:garbage_collect(Pid, [{async,Ref}])

the reply message could in certain circumstances be
lost.

The problem is that cleanup_sys_tasks() is never called
if there are dirty tasks, but no other active system tasks.

Also shorten the long waiting times in the test case
binary_SUITE:trapping/1 to make it much more likely that
the GC have not already finished when the process is killed.