stdlib: Improve edlin handling of unicode chars
Let edlin handle grapheme clusters instead of codepoints to
improve the handling multi-codepoints characters.

The ttsl driver (and protocol) still expects all lengths as
codepoints.

Previously it was expected that each codepoint used (at least) one
terminal column for each codepoint, and a hack was made for wide
characters (multicolumn) by patching in TAGGED characters to occupy
the extra space so that codepoint index was equal column index.

This didn't work at all for combining codepoints that do not occupy any
more space than the previous character.

Improved this handling by calculating column positions in move_cursor.
This is based on wcwidth() and is not perfect, wcwidth() is wrong for
some codepoints and wcwidth() can not know with Hangul graphemes for
example.  But it works better than before without making a major change
in the protocol.