Accept IPv6 address %suffixes when parsing
This only implements decimal suffixes, and furthermore only with
a leading "0".  A complete implementation would probably allow one
and probably two digit suffixes without.

But what primarily is missing is translating interface names to
and from interface indexes.

This also only implements the parsing, and uses the FreeBSD trick of
squeezing in the Scope Id in the second 16-bit word of the fe80::/64
or ff02:/64 address prefix.  But inet_drv is not prepared to handle
this word, so it might only work on FreeBSD, not being supported even
there...  So inet_drv needs to handle this too.