dtls: Make HelloVerifyRequest version adhere to RFC
ERL-434

RFC6347 says about hello_verify_request version field as follow

https://tools.ietf.org/html/rfc6347#page-16

The server_version field has the same syntax as in TLS. However, in
order to avoid the requirement to do version negotiation in the
initial handshake, DTLS 1.2 server implementations SHOULD use DTLS
version 1.0 regardless of the version of TLS that is expected to be
negotiated.

But current DTLS server responses DTLS1.2 instead of DTLS1.0.