erts: Fix bug in quick alloc
The effect of the race is that a pre-allocated memory block
is inserted last without updating tail.data.last, which will cause
all subsequent insertions to also fail to update tail.data.last.

Hence all pre-allocation for this quick alloc instance is leaked
for this thread and will fallback on erts_alloc.