Introduce new_small_map_lit op
Take advantage of the fact that small maps have a tuple for keys.
When new map is constructed and all keys are literals, we can construct
the entire keys tuple as a literal.

This should reduce the memory of maps created with literal keys almost by half,
since they all can share the same keys tuple.