crypto: pk sign with new function from PR838
The function  pkey_verify_nif handles pk signing with a general approach that is enabled by EVP api in OpenSSL 1.0.0
The rejected PR838 introduced that function and lots of other stuff so far neglected. It also made some different rsa padding methods possible (included).

Since the crypto code base has changed significantly, it was an overhelming work to try a git merge.  Therefore this commit is a manual move of the source code from the PR into current maint.

This commit concentrates to use the new function compatible with the old functions. This includes some #if:s for different versions, compatibility code for 0.9.8.

Lacking:  test cases, exporting the more general erlang api, documentation...