Once again include get_map_elements in blocks
This was previously done, but it was found to be unsafe and first modified in
c2035ebb8b7bbdeee1ff03a348ba39edef1050b1 and then fully removed in
0c599bcad1e7f5f66dd2342ab27791048145e892.

Including get_map_elements in blocks is problematic because it can jump out
of the block and potentially pollute some of the target registers.

We solve this problem by only ever allowing the instruction to be the
first element of a block and by keeping track of "protected" registers when
optimising move instructions - in case of multi-key get_map_elements, we
consider the source register protected. Finally we perform proper liveness
analysis over the instruction when it's included as part of a block in
`beam_utils:live_opt`.

The changes pass all the tests introduced as failure cases when the optimisation
was initially limited and removed.