Limit SCTP testing in traffic suite
All SCTP testing on Solaris/Sparc was disabled in commit 69c5a741, but
there are issues on other platforms as well, resulting in failures in
diameter_gen_sctp_SUITE and more. The problems seem to be load-related,
when testcases are run in parallel, so only run a smallish subset
sequentially until the gen_sctp suite runs without error.