Enhance elimination of useless tests in beam_jump
It can happen we have the following situation:

    {test,is_tuple,Fail,[R1]}
    {test,test_arity,Fail,[R1,N1]}
    {get_tuple_element,R1,N2,R2}
    {test,is_eq_exaqct,Fail,[R2,Atom]}
    {jump,Fail}

Previously, the optimisation would eliminate the last is_eq_exact test, but
we can do more. If the register R2 is not used in Fail, we can eliminate the
get_tuple_element instruction as well as all the preceding tests. Ultimately,
the whole sequence can be replaced by:

    {jump,Fail}