Fix ANSI support in the console
The ANSI support doesn't work properly with edlin, the issue can be noticed when you try to use the history of the shell and the prompt prefix has ANSI (https://github.com/elixir-lang/elixir/issues/6448). The problem is that when a `\e` character appears, it handles it like a new line, dropping the buffer before it.

The solution is to always add the `\e` to the buffer like a regular character and handle it when writing the buffer instead.