beam_makeops: Introduce the new type 'W' (machine word)
As a preparation for potentially improving packing in the future,
we will need to make sure that packable types have a defined maximum
size.

The packer algorithm assumes that two 'I' operands can be packed
into one 64-bit word, but there are instructions that use an 'I'
operand to store a pointer. It only works because those instructions
are not packed for other reasons.

Introduce the 'W' type and use it for operands that don't fit in
32 bits.