Fix disksup:get_disksup_data() for High Sierra
On macOS 10.13 (High Sierra), disksup cannot grab information
for any disks that use the new APFS file system. The reason
is that the invocation of /bin/df explicitly limits the
type of file systems to show:

    /bin/df -i -k -t ufs,hfs

Add 'apfs' to also look at disks with APFS:

    /bin/df -i -k -t ufs,hfs,apfs

That also works on older versions of macOS.

An alternative solution would be to use the -l flag to only
show local file systems. However, that would mean that USB
memory sticks would also show up, and it might not be desirable
to monitor them.

https://bugs.erlang.org/browse/ERL-461