ssl: Mend Max session handling
The commit 256e01ce80b3aadd63f303b9bda5722ad313220f was a misunderstanding
that actually broke the implementation.

It is not so important to keep specific max, rather max is a threshold
when the table should be shrinked as to not grow indefinitely.

New sessions are created when the id is created and may be short lived
it they are not registered for reuse due to handshake failure.