Remove backward compatibility support for float literals
Starting from R16B, floating point constants are stored
in the literal pool.

The backward compatibility support for the old representation
of floats has been kept long enough. Also, we might want to
ensure that all literals are unique in the future, and we certainly
don't want to update this code to ensure uniqueness.